+++ tests/testthat/test-receptiviti.R
@@ -2,7 +2,7 @@ options(stringsAsFactors = FALSE)
 key <- Sys.getenv("RECEPTIVITI_KEY")
 secret <- Sys.getenv("RECEPTIVITI_SECRET")
 text <- "a text to score"
-temp <- tempdir()
+temp <- normalizePath(tempdir(), "/")
 temp_cache <- paste0(temp, "/temp_cache")
 Sys.setenv(RECEPTIVITI_KEY = 123, RECEPTIVITI_SECRET = 123)
 on.exit(Sys.setenv(RECEPTIVITI_KEY = key, RECEPTIVITI_SECRET = secret))
@@ -248,8 +248,10 @@ test_that("reading from files works", {
   writeLines(texts[1], paste0(temp_source, "0.txt"))
 
   txt_directory <- receptiviti(temp_source, cache = temp_cache, bundle_size = 25)
+  initial <- cbind(id = files_txt, initial)
   expect_true(all(initial$text_hash %in% txt_directory$text_hash))
   expect_false(anyDuplicated(txt_directory$text_hash) == 0)
+  txt_directory <- txt_directory[-1, ]
   txt_directory <- txt_directory[!duplicated(txt_directory$text_hash), ]
   rownames(txt_directory) <- txt_directory$text_hash
   txt_directory <- txt_directory[initial$text_hash, ]
@@ -268,15 +270,16 @@ test_that("reading from files works", {
   rownames(csv_directory) <- csv_directory$text_hash
   csv_directory <- csv_directory[initial$text_hash, ]
   rownames(csv_directory) <- NULL
+  initial$id <- files_csv
   expect_equal(csv_directory, initial)
 
-  expect_equal(receptiviti(file_txt, cache = temp_cache, in_memory = FALSE)[, -1], initial)
+  expect_equal(receptiviti(file_txt, cache = temp_cache, in_memory = FALSE)[, -1], initial[, -1])
   expect_equal(
-    receptiviti(file_txt, collapse_lines = TRUE, cache = temp_cache),
+    receptiviti(file_txt, collapse_lines = TRUE, cache = temp_cache)[, -1],
     receptiviti(paste(texts, collapse = " "), cache = temp_cache)
   )
   expect_error(receptiviti(file_csv, cache = temp_cache))
-  expect_equal(receptiviti(file_csv, text_column = "text", cache = temp_cache)[, -1], initial)
+  expect_equal(receptiviti(file_csv, text_column = "text", cache = temp_cache)[, -1], initial[, -1])
 
   alt_id <- receptiviti(file_csv, text_column = "text", id_column = "id", cache = temp_cache)
   expect_identical(alt_id, receptiviti(csv_data, text_column = "text", id = "id", cache = temp_cache))